feat: Move ContextKey from public API to internal API
