+++ dd-trace-core/src/main/java/datadog/trace/core/scopemanager/ScopeContext.java
@@ -1,12 +1,12 @@
 package datadog.trace.core.scopemanager;
 
-import static datadog.trace.context.ContextKey.named;
+import static datadog.trace.bootstrap.instrumentation.api.ContextKey.named;
 
 import datadog.trace.bootstrap.instrumentation.api.Baggage;
 import datadog.trace.bootstrap.instrumentation.api.AgentScopeContext;
 import datadog.trace.bootstrap.instrumentation.api.AgentScopeManager;
 import datadog.trace.bootstrap.instrumentation.api.AgentSpan;
-import datadog.trace.context.ContextKey;
+import datadog.trace.bootstrap.instrumentation.api.ContextKey;
 import java.util.Arrays;
 import java.util.Objects;
